Display bugs
-Fixed Hungary Huszair display bug
Overrides UnitPanel_XP2, english loc included

Mechanics:
Added the baseline modifier that adds +0 to huszair. Attached a modifier string so it gets displayed by ui.
Override UnitPanel UI to show total amount of CS and concise list of citystates. Having +2 in each line and 50 lines of localized text would be very annoying for all.
Extra:
BBG already had Unit Panel override (by JackTheNarrator) that was not active. It had condemn heretic mechanic. I renamed his override to unitpanel_bbg_legacy, so it still hangs around like it used to and new override is called UnitPanel_BBG.lua and is activated in the modinfo

Potential Problem: BSM overrides the same panel, need to add action criteria in BSM that goes like if BBG5 active overide the BBG5 panel instead of XP2 panel like it is now. So expect new BSM for this conflict resolution and fix of the Writing boost with spectator.

Fixed Portugal Trade Routs Preview
Replaces the support script TradeSupport.lua. It is actually a Firaxis indexing bug C++ indexing starts with 0, while lua starts with 1(simillar to no available trade routs from capital bug fixed by Vodario). TradeSupport is updated with the same name, so it works clean with all the other mods (because firaxis engine loads the last queued file with the same name). No other of our mods replaces this file, so it is 100% safe.

Modinfo changed to 5.1.2
and respective necessary lines added for both fixes.

P.S. Didn't want to separate these 2 PR's to not make mistakes in modinfo and risk having the error: One or More Mods Failed to load.